fix(tracker): clear the lint findings, and the two bugs behind them
CI's golangci-lint job is red with 176 findings and half of them are this
package's. These are the ones that are not doc comments or err shadowing,
and two of them are real defects the style linters happened to be
standing next to.

A NO-OP REPLACED A STRING WITH ITSELF at the head of the `all` /
`not_all` field filter — `strings.Replace(clause, "SELECT v.task_id",
"SELECT v.task_id", 1)` — left over from a shape that widened the select
list. It read as though the projection had to change for the count to
work, which it does not: `COUNT(DISTINCT …)` in the HAVING reads the
column directly. The arm had NO TEST AT ALL, which is why a line that did
nothing sat in it. It has two now: that the grouping lands INSIDE the
subquery's own parentheses (outside, the statement does not parse), that
the count is DISTINCT (without it a task holding one option twice
satisfies a two-option filter), that the set's size is the last argument
bound, and that `not_all` negates the whole membership rather than
comparing row-wise. Both mutated red.

AN APPEND WROTE THROUGH ANOTHER SLICE'S CAPACITY in the view listing:
`listing.Views = append(implicit, saved...)`. Safe today only because
nothing reads `implicit` afterwards — which is a property of the current
line ordering rather than of the code, and the kind that stops holding
the day somebody adds a read above it.

The rest:

  * Three `nilerr` sites are deliberate and each already carried its
    reason in prose — the applier salvaging a value it cannot decode, a
    scope from a newer peer widening to the domain, a claim failing open.
    Each gets a //nolint naming the invariant, which is this repository's
    own convention and what its 134 existing directives look like.
  * `ParseQuery`'s doc comment had drifted onto the type declared beneath
    it, so revive reported the function as undocumented and the type's
    own comment as a stray paragraph. Moved back.
  * `nullableRaw` lost its last caller when migration 0010 dropped the
    catalogue tables; `explodeTask`'s `c` and `prioritisedWake`'s `at`
    are parameters nothing reads.
  * Two test helpers asserted on an error's type directly, which misses a
    wrapped one — the exact failure `errors.As` exists for, in helpers
    named `asGate` and `asFuture`.
  * A deferred `w.Close()` unchecked, three `OpKind` conversions of a
    value that is already one, a `!(a < b)` that reads as the negation of
    an ordering nobody wrote, a literal control character, and `recieved`.

+++ internal/tracker/write.go
@@ -423,7 +423,7 @@ func (w *Writer) UpdateTask(ctx context.Context, opID, id, project string,
 					"removed by %s at %s; restore it first",
 					id, current.Removed.By, current.Removed.At.Format(time.RFC3339))
 			}
-			if ifMatch != 0 && uint64(current.Version) != ifMatch {
+			if ifMatch != 0 && current.Version != ifMatch {
 				return statelog.Decision{}, fmt.Errorf("%w: task %s is at "+
 					"version %d and the edit was conditioned on %d — re-read "+
 					"it and decide again rather than re-sending this patch",
